Create onlineshoppingdatabase_project
The aim of this project is to create and execute a thorough online marketplace platform, utilizing an Oracle database, to empower our customer to rival industry leaders such as Amazon and Walmart. Our team will prioritize the creation of a refined database structure, the development of crucial tables, and the designing of efficient PL/SQL stored procedures. The operations would encompass essential features, including product management, storage of client information, processing of orders, generation of invoices, and provision of individualized product suggestions. Furthermore, the project seeks to provide comprehensive reports that offer valuable information for managerial decision-making. These reports will offer a comprehensive overview of the business's performance during specified time periods and provide in-depth analysis of product sales. Our objective is to provide our customer with an innovative online marketplace that distinguishes itself in the competitive market through careful design, testing, and implementation